1) Added getTracksInGroup() method. 2) Removed unnecessary leader-result files. 3) Minor code cleanup.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@2176 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/world.cpp
@@ -254,7 +254,7 @@ void World::resetAllKarts()
                 if(!material)
                 {
                     fprintf(stderr, "ERROR: no valid starting position for kart %d on track %s.\n",
-                        i-m_kart.begin(), m_track->getIdent().c_str());
+			    (int)(i-m_kart.begin()), m_track->getIdent().c_str());
                     exit(-1);
                 }
                 all_finished=false;
